requirements.txt
pytest~=7.1.3
allure-pytest
selene>=2.0.b8
selenium~=4.4.3
python-dotenv~=0.21.0
pycurl
requests
voluptuous
pytest-voluptuous